third_party/WebKit/LayoutTests/fast/dom/MutationObserver/observe-options-attributes-expected.txt

Test MutationObserver.observe options related to attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS records.length is 1
PASS records[0].oldValue is "1"
PASS records.length is 1
PASS records[0].oldValue is null
PASS records.length is 1
PASS records[0].oldValue is null
PASS successfullyParsed is true

TEST COMPLETE
